K19DKPA-60 manually Tested "find all", pagination and "find by id" A lot of bugs fixed. DTOs for objects nested in bundle required.

+++ backend/src/main/java/edu/netcracker/backend/model/Service.java
@@ -15,7 +15,7 @@ public class Service {
 
     @PrimaryKey("service_id")
     @EqualsAndHashCode.Include
-    private Long serviceId;
+    private Integer serviceId;
 
     @Attribute("service_name")
     private String serviceName;
